code refactor for response body
code refactor for response body

+++ src/main/java/com/productcatalog/app/service/impl/ProductServiceImpl.java
@@ -16,6 +16,7 @@ import com.productcatalog.app.model.Status;
 import com.productcatalog.app.repository.ApprovalQueueRepository;
 import com.productcatalog.app.repository.ProductRepository;
 import com.productcatalog.app.response.ProductCatalogConstants;
+import com.productcatalog.app.response.ResponseHandler;
 import com.productcatalog.app.service.ProductService;
 
 import lombok.extern.log4j.Log4j2;
@@ -29,29 +30,11 @@ public class ProductServiceImpl implements ProductService {
 	@Autowired
 	private ApprovalQueueRepository approvalQueueRepository;
 
-	@Override
-	public List<Product> searchProductsBasedOnSearchCriteria(String productName, Double minPrice, Double maxPrice,
-			LocalDateTime minPostedDate, LocalDateTime maxPostedDate) {
-		// Validations: Ensure that the maxPrice is greater than or equal to minPrice
-		if (minPrice != null && maxPrice != null && minPrice.compareTo(maxPrice) >= 0) {
-			throw new IllegalArgumentException("maxPrice should be greater than or equal to minPrice");
-		}
-
-		// Validations: Ensure that the maxPostedDate is after or equal to minPostedDate
-		if (minPostedDate != null && maxPostedDate != null && minPostedDate.isAfter(maxPostedDate)) {
-			throw new IllegalArgumentException("maxPostedDate should be after or equal to minPostedDate");
-		}
-		log.info("After ValidationRetrieving product on the basis of the search criteria");
-		return productRepository.findByNameEqualsIgnoreCaseOrPriceBetweenOrPostedDateBetween(productName, minPrice,
-				maxPrice, minPostedDate, maxPostedDate);
-	}
-
-	public ResponseEntity<String> createProductwithApprovalCheck(Product product) {
+	public ResponseEntity<Object> createProductwithApprovalCheck(Product product) {
 		String response = null;
 		if (product.getPrice() <= ProductCatalogConstants.MAX_PRICE) {
 
 			if (product.getPrice() > 5000) {
-
 				ApprovalQueue approvalQueue = new ApprovalQueue();
 				product.setPostedDate(LocalDateTime.now());
 				approvalQueue.setName(product.getName());
@@ -60,30 +43,32 @@ public class ProductServiceImpl implements ProductService {
 				approvalQueue.setStatus(product.getStatus());
 				approvalQueueRepository.save(approvalQueue);
 				log.info("Creating product with price>5000, Hence added to approval queue");
-				response = "Product Added To Approval Queue";
+				response = "Product Added To Approval Queue as price is more than 5000";
 
 			} else {
 				productRepository.save(product);
 				log.info("Product Created successfully");
 				response = "Product Created Successfully";
 			}
-			return new ResponseEntity<>(response, HttpStatus.OK);
+			return ResponseHandler.generateResponse(response, HttpStatus.OK);
+
 
 		} else {
-
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Product price exceeds 10,000. Hence not saved.");
+			
+				throw new IllegalArgumentException("Product price exceeds 10,000. Hence not saved.");
+			
 		}
 	}
 
-	public ResponseEntity<String> updateProductWithApprovalCheck(Long productId, Product updatedProduct) {
+	public ResponseEntity<Object> updateProductWithApprovalCheck(Long productId, Product updatedProduct) {
 		Optional<Product> productOptional = productRepository.findById(productId);
 		if (productOptional.isPresent()) {
 			Product product = productOptional.get();
 			Double previousPrice = product.getPrice();
 
 			if (updatedProduct.getPrice() > 10000) {
-				return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-						.body("Product price exceeds 10,000. Hence not updated.");
+				throw new IllegalArgumentException("Product price exceeds 10,000. Hence not updated.");
+				
 			}
 			// Check if the new price is more than 50% of the previous price
 			Double fiftyPercentOfPreviousPrice = previousPrice * 0.5;
@@ -98,9 +83,8 @@ public class ProductServiceImpl implements ProductService {
 				approvalQueueRepository.save(approvalQueue);
 				log.info("Product price updated to 50% more than previous price, hence added to approval queue");
 				productRepository.save(product);
-
-				return ResponseEntity.ok()
-						.body("Product Sent for Approval as price is higher than 50% of previous value");
+				return ResponseHandler.generateResponse("Product Sent for Approval as price is higher than 50% of previous value", HttpStatus.OK);
+				
 			} else {
 				product.setName(updatedProduct.getName());
 				product.setPrice(updatedProduct.getPrice());
@@ -108,16 +92,18 @@ public class ProductServiceImpl implements ProductService {
 				product.setPostedDate(LocalDateTime.now());
 				log.info("Product Saved Successfully");
 				productRepository.save(product);
-				return new ResponseEntity<>("Product Updated Successfully", HttpStatus.OK);
+				return ResponseHandler.generateResponse("Product Updated Successfully", HttpStatus.OK);
+
+				
 			}
 		} else {
-			throw new ResourceNotFoundException("No oduct not found with ID: " + productId + "found");
+			throw new ResourceNotFoundException("No product not found with ID: " + productId + "found");
 		}
 	}
 
-	public ResponseEntity<String> deleteProductWithApproval(Long productId) {
+	public ResponseEntity<Object> deleteProductWithApproval(Long productId) {
 		Product prod = productRepository.findById(productId)
-				.orElseThrow(() -> new ResourceNotFoundException("No Product with ID :" + productId + " found!"));
+				.orElseThrow(() -> new ResourceNotFoundException("No Product with ID " + productId + " found!"));
 		ApprovalQueue approvalQueue = new ApprovalQueue();
 		prod.setStatus(Status.INACTIVE);
 		approvalQueue.setProductId(productId);
@@ -128,13 +114,17 @@ public class ProductServiceImpl implements ProductService {
 		log.info("Product requested to be deleted, Hence added to approval queue");
 		approvalQueueRepository.save(approvalQueue);
 		productRepository.save(prod);
-		return new ResponseEntity<>("Product Deleted successfully", HttpStatus.OK);
+		return ResponseHandler.generateResponse("Product Deleted successfully", HttpStatus.OK);
+		
 
 	}
 
-	public ResponseEntity<String> approveProduct(Long approvalId) {
+	public ResponseEntity<Object> approveProduct(Long approvalId) {
+		if (approvalId == null) {
+			throw new IllegalArgumentException("Approval ID cannot be null");
+		}
 		ApprovalQueue approveQueueData = approvalQueueRepository.findById(approvalId).orElseThrow(
-				() -> new ResourceNotFoundException("No ID in Approval Queue with ID :" + approvalId + " found!"));
+				() -> new ResourceNotFoundException("No ID in Approval Queue with ID " + approvalId + " found!"));
 		Long productId = approveQueueData.getProductId();
 		if (productId != null) {
 			Optional<Product> productOptional = productRepository.findById(productId);
@@ -158,15 +148,39 @@ public class ProductServiceImpl implements ProductService {
 			productRepository.save(product);
 		}
 		approvalQueueRepository.delete(approveQueueData);
-		return new ResponseEntity<>("Product approved successfully and product updated", HttpStatus.OK);
+		return ResponseHandler.generateResponse("Product approved successfully and product updated", HttpStatus.OK);
+		
 	}
 
-	public ResponseEntity<String> rejectProduct(Long approvalId) {
+	public ResponseEntity<Object>rejectProduct(Long approvalId) {
+		if (approvalId == null) {
+			throw new IllegalArgumentException("Approval ID cannot be null");
+		}
 		ApprovalQueue approveQueueData = approvalQueueRepository.findById(approvalId).orElseThrow(
-				() -> new ResourceNotFoundException("No ID in Approval Queue with ID :" + approvalId + " found!"));
+				() -> new ResourceNotFoundException("No ID in Approval Queue with ID " + approvalId + " found!"));
 		log.info("Product rejected. Product state unchanged");
 		approvalQueueRepository.delete(approveQueueData);
-		return new ResponseEntity<>("Product Rejected", HttpStatus.OK);
+		return ResponseHandler.generateResponse("Product Rejected Successfully", HttpStatus.OK);
+
 
 	}
+
+	@Override
+	public List<Product> searchProductsBasedOnSearchCriteria(String productName, Double minPrice, Double maxPrice,
+			LocalDateTime minPostedDate, LocalDateTime maxPostedDate) throws Exception{
+		// Validations: Ensure that the maxPrice is greater than or equal to minPrice
+		if (minPrice != null && maxPrice != null && minPrice.compareTo(maxPrice) >= 0) {
+			throw new IllegalArgumentException("maxPrice should be greater than or equal to minPrice");
+		}
+
+		// Validations: Ensure that the maxPostedDate is after or equal to minPostedDate
+		if (minPostedDate != null && maxPostedDate != null && minPostedDate.isAfter(maxPostedDate)) {
+			throw new IllegalArgumentException("maxPostedDate should be after or equal to minPostedDate");
+		}
+		log.info("After ValidationRetrieving product on the basis of the search criteria");
+		return productRepository.findByNameEqualsIgnoreCaseOrPriceBetweenOrPostedDateBetween(productName, minPrice,
+				maxPrice, minPostedDate, maxPostedDate);
+	}
+	
+	
 }
